# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Templet Webinars API",
        "description": "API de solo lectura para consultar submissions, asistentes y estadísticas de los webinars. Autenticación por API key en el header `X-Api-Key`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor actual"
        }
    ],
    "paths": {
...
        "/api/v1/submissions": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Lista los registros (submissions) de los webinars",
                "description": "Devuelve los registros capturados en las landings de webinars, con filtros por cliente, webinar, campaña de Salesforce, fechas y UTM.",
                "operationId": "listSubmissions",
                "parameters": [
                    {
                        "name": "client",
                        "in": "query",
                        "description": "Slug o ID del cliente",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "libertynet"
                        }
                    },
                    {
                        "name": "webinar_id",
                        "in": "query",
                        "description": "ID del webinar",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "webinar_slug",
                        "in": "query",
                        "description": "Slug del webinar",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "campaign",
                        "in": "query",
                        "description": "Campaign Salesforce Field del webinar (campo `campaign`). Usado para enlazar con la campaña en Salesforce.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "7013h000000abcAAA"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha inicial (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha final (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "utm_source",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "utm_campaign",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Busca dentro del JSON `data` (nombre, email, etc.)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "registered_in_zoom",
                        "in": "query",
                        "description": "true = solo los registrados en Zoom, false = solo los no registrados",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Resultados por página (1-200, default 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado de submissions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Submission"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "API key faltante o inválida"
                    },
                    "403": {
                        "description": "La API key no tiene acceso a ese cliente"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/submissions/{id}": {
            "get": {
                "tags": [
                    "Submissions"
                ],
                "summary": "Detalle de un registro",
                "operationId": "showSubmission",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Submission"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Submission": {
                "title": "Submission",
                "description": "Registro capturado en la landing de un webinar.",
                "properties": {
                    "id": {
                        "description": "Definiciones de los objetos que devuelve la API, solo para la documentación\nSwagger. Esta clase no se instancia en runtime.",
                        "type": "integer",
                        "example": 1421
                    },
                    "webinar": {
                        "$ref": "#/components/schemas/WebinarRef"
                    },
                    "data": {
                        "description": "Campos dinámicos definidos en el form_schema del webinar.",
                        "type": "object",
                        "example": {
                            "name": "Ana",
                            "last_name": "Pérez",
                            "email": "ana@empresa.com",
                            "phone": "+525512345678",
                            "company": "Empresa SA"
                        },
                        "additionalProperties": true
                    },
                    "utm": {
                        "properties": {
                            "source": {
                                "type": "string",
                                "example": "paid",
                                "nullable": true
                            },
                            "medium": {
                                "type": "string",
                                "example": "facebook",
                                "nullable": true
                            },
                            "campaign": {
                                "type": "string",
                                "nullable": true
                            },
                            "term": {
                                "type": "string",
                                "nullable": true
                            },
                            "content": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "sent_to_clay_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "registered_in_zoom_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "WebinarRef": {
                "title": "WebinarRef",
                "description": "Referencia al webinar y su cliente.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "title": {
                        "type": "string",
                        "example": "LibertyNet SASE"
                    },
                    "slug": {
                        "type": "string",
                        "example": "libertynet-sase"
                    },
                    "campaign_salesforce_field": {
                        "description": "Campo `campaign` del webinar. Es el identificador con el que se enlaza la campaña en Salesforce.",
                        "type": "string",
                        "example": "7013h000000abcAAA",
                        "nullable": true
                    },
                    "client": {
                        "$ref": "#/components/schemas/ClientRef"
                    }
                },
                "type": "object"
            },
            "ClientRef": {
                "title": "ClientRef",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "LibertyNet"
                    },
                    "slug": {
                        "type": "string",
                        "example": "libertynet"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API key generada con `php artisan api-key:create`.",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
...
}